benchmark/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depth-bench"
version = "0.1.0"
description = "Unified real-time benchmarking harness for depth-estimation models on Intel RealSense D435i."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "opencv-python",
    "pyyaml",
    "pandas",
    "tabulate",
    "matplotlib",
]

[project.optional-dependencies]
camera = ["pyrealsense2"]
viz = ["open3d"]

[tool.setuptools]
packages = ["benchmark", "benchmark.camera", "benchmark.io", "benchmark.adapters", "benchmark.runners", "benchmark.metrics", "benchmark.viz"]
